cleaning up error handling, allow subsequent auth flows
