+++ auth-lobby/src/LoginPage.tsx
@@ -11,24 +11,37 @@ interface Props {
 
 function LoginPage(props: Props) {
   const [pin, setPin] = useState<number | null>(null)
+  const [searching, setSearching] = useState<boolean>(false)
+  const [error, setError] = useState<boolean>(false)
 
   const linkDevice = async () => {
-    const requester = await awake.Requester.create(
-      env.RELAY_HOST,
-      env.ROOT_USER,
-      await props.authStore.getDid(),
-    )
-    const pin = await requester.findProvider()
-    setPin(pin)
-    const token = await requester.awaitDelegation()
-    await props.authStore.addUcan(token)
-    console.log('SUCCESS')
-    props.checkAuthorized()
+    setError(false)
+    try {
+      const requester = await awake.Requester.create(
+        env.RELAY_HOST,
+        env.ROOT_USER,
+        await props.authStore.getDid(),
+      )
+      setSearching(true)
+      const pin = await requester.findProvider()
+      setSearching(false)
+      setPin(pin)
+      const token = await requester.awaitDelegation()
+      await props.authStore.addUcan(token)
+      requester.close()
+      console.log('SUCCESS')
+      props.checkAuthorized()
+    } catch (_err) {
+      setPin(null)
+      setError(true)
+    }
   }
 
   return (
     <div>
-      <button onClick={linkDevice}>Link Device</button>
+      {error && <div>Linking failed. Try again?</div>}
+      {!searching && !pin && <button onClick={linkDevice}>Link Device</button>}
+      {searching && <div>Searching for provider...</div>}
       {pin && <div>Pin: {pin}</div>}
     </div>
   )
