+++ ws-relay/README.md
@@ -1,0 +1,5 @@
+# Web socket relay
+
+A simple topic-based web sockets relay server.
+
+Messages that are sent to a room are broadcasted to all other participants in that room
