--- ws-relay/ava.config.js
@@ -1,6 +1,0 @@
-export default {
-  extensions: {
-    ts: 'module',
-  },
-  nodeArguments: ['--loader=ts-node/esm'],
-}
