+++ ws-relay/src/bin.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import server from './server.js'
+import server from './index.js'
 
 const envPort = process.env.PORT
 const port = envPort ? parseInt(envPort) : 3005
